smartbudgeting-api/requirements.txt
Flask==3.1.0
google-cloud-storage==2.19.0
pandas==2.1.3
scikit-learn==1.4.0